[BUGFIX] Trigger submit of RSA encrypted form properly
The change of #76120 changed the selector of possible submit buttons.
However, in Firefox a submit is not triggered when a focussed submit
element is clicked via JavaScript.

The code is changed to enforce a submit of the form instead.

RsaEncryptionWithLib.min.js created with:
../../../../../../Build/node_modules/uglify-js/bin/uglifyjs \
RsaLibrary.js RsaEncryption.js > RsaEncryptionWithLib.min.js

Resolves: #84253
Related: #76120
Releases: master, 8.7, 7.6
Change-Id: Id30f4536dccec28edc0d909114c8fde71614bc30
Reviewed-on: https://review.typo3.org/56143
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Kevin Ditscheid <kevinditscheid@gmail.com>
Tested-by: Kevin Ditscheid <kevinditscheid@gmail.com>
Reviewed-by: Andreas Wolf <andreas.wolf@typo3.org>
Tested-by: Andreas Wolf <andreas.wolf@typo3.org>