[FEATURE] Allow TCA description property
When the site configuration module has been introduced, it came
with a custom functionality to show an additional help text
when editing site records between the field label and the field input.

This useful feature is now changed into a general TCA feature
available everywhere: A new field information node expansion / "wizard"
is added to all form elements, the inline and flex containers: If
the property "description" is set for a TCA column type (same array
level as "label", it will show the value as localized string between
the field label and the input section.

There are three available render types for "wizard a-like" output:
* Field information - text between label+field
* Field control - buttons next to input sections like the link popup button
* Field wizards - clickable stuff below the input section, for example
  the localization state selector
If a field has been set to readOnly=true in TCA, field control and field
wizards do not make sense to render since they are meant to act with the
field value.
The field information node however has only informational character
which is useful for readOnly fields, too. Thus, this node expansion
type is now the only one that is always rendered, even if a field has
been set to readOnly.

Note this patch is fully covered by ext:styleguide (master) to have
examples for all changed elements now using the description property.

Resolves: #85410
Releases: master
Change-Id: Idcfacafa19b8208614b653b8fac22ce47bca3b8f
Reviewed-on: https://review.typo3.org/57397
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Andreas Wolf <andreas.wolf@typo3.org>
Tested-by: Andreas Wolf <andreas.wolf@typo3.org>